+++ application/modules/editor/Services/OpenTM2/HttpApi.php
@@ -34,6 +34,8 @@ use MittagQI\Translate5\T5Memory\Enum\StripFramingTags;
  */
 class editor_Services_OpenTM2_HttpApi extends editor_Services_Connector_HttpApiAbstract
 {
+    private const DATE_FORMAT = 'Ymd\THis\Z';
+
     const MAX_STR_LENGTH = 2048;
 
     /**
@@ -409,7 +411,8 @@ class editor_Services_OpenTM2_HttpApi extends editor_Services_Connector_HttpApiA
         editor_Models_Segment $segment,
         string $filename,
         string $tmName,
-        bool $save2disk = true
+        bool $save2disk = true,
+        bool $useSegmentTimestamp = false
     ): bool {
         $this->error = null;
 
@@ -420,9 +423,13 @@ class editor_Services_OpenTM2_HttpApi extends editor_Services_Connector_HttpApiA
             return false;
         }
 
+        $timestamp = $useSegmentTimestamp
+            ? (new DateTimeImmutable($segment->getTimestamp()))->format(self::DATE_FORMAT)
+            : $this->nowDate();
+
         $json->documentName = $filename; // 101 doc match
         $json->author = $segment->getUserName();
-        $json->timeStamp = $this->nowDate();
+        $json->timeStamp = $timestamp;
         $json->context = $segment->getMid(); //INFO: this is segment stuff
         // t5memory does not understand boolean parameters, so we have to convert them to 0/1
         $json->save2disk = $save2disk ? '1' : '0';
@@ -569,7 +576,7 @@ class editor_Services_OpenTM2_HttpApi extends editor_Services_Connector_HttpApiA
      */
     protected function nowDate()
     {
-        return gmdate('Ymd\THis\Z');
+        return gmdate(self::DATE_FORMAT);
     }
 
     /**